+++ packages/parse-tmx/src/util/parseNode.ts
@@ -6,7 +6,7 @@ import type { TMXLayerParsed } from "@/models/tmx/parsed/TMXLayerParsed";
 import { parseGroup } from "@/util/parseGroup";
 import { parseLayer } from "@/util/parseLayer";
 import { parseTileLayer } from "@/util/parseTileLayer";
-import { InvalidOperationError, Operation } from "esposter-shared";
+import { InvalidOperationError, Operation } from "@esposter/shared";
 
 export const parseNode = <
   TNode extends TMXLayerNode | TMXGroupLayerNode,