+++ server/trpc/routers/post.ts
@@ -10,8 +10,8 @@ import { ranking } from "@/services/post/ranking";
 import { getCursorPaginationData } from "@/services/shared/pagination/cursor/getCursorPaginationData";
 import { getCursorWhere } from "@/services/shared/pagination/cursor/getCursorWhere";
 import { parseSortByToSql } from "@/services/shared/pagination/sorting/parseSortByToSql";
+import { NotFoundError } from "@esposter/shared";
 import { and, eq, isNotNull, isNull } from "drizzle-orm";
-import { NotFoundError } from "esposter-shared";
 import { z } from "zod";
 
 const readPostInputSchema = selectPostSchema.shape.id;