+++ services/dungeons/area/getEncounterArea.ts
@@ -1,6 +1,6 @@
 import { encounterAreas } from "@/assets/dungeons/data/encounterAreas";
 import type { Area } from "@/generated/tiled/propertyTypes/enum/Area";
-import { NotFoundError } from "esposter-shared";
+import { NotFoundError } from "@esposter/shared";
 
 export const getEncounterArea = (area: Area) => {
   const encounterArea = encounterAreas.find((a) => a.id === area);