+++ store/dungeons/settings/volume.ts
@@ -3,7 +3,7 @@ import { SettingsOption } from "@/models/dungeons/scene/settings/SettingsOption"
 import { dayjs } from "@/services/dayjs";
 import { useSettingsStore } from "@/store/dungeons/settings";
 import { step } from "@/util/math/ease/step";
-import { InvalidOperationError, Operation } from "esposter-shared";
+import { InvalidOperationError, Operation } from "@esposter/shared";
 import { Direction } from "grid-engine";
 import type Slider from "phaser3-rex-plugins/plugins/slider";