+++ store/tableEditor/index.test.ts
@@ -2,7 +2,7 @@ import { TodoListItem } from "@/models/tableEditor/todoList/TodoListItem";
 import { ITEM_ID_QUERY_PARAM_KEY } from "@/services/tableEditor/constants";
 import { useTableEditorStore } from "@/store/tableEditor";
 import { useItemStore } from "@/store/tableEditor/item";
-import { InvalidOperationError, Operation } from "esposter-shared";
+import { InvalidOperationError, Operation } from "@esposter/shared";
 import { createPinia, setActivePinia } from "pinia";
 import { beforeEach, describe, expect, test } from "vitest";