+++ packages/parse-tmx/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.30](https://github.com/Esposter/Esposter/compare/v1.0.28...v1.0.30) (2024-05-18)
+
+**Note:** Version bump only for package parse-tmx
+
+
+
+
+
+## [1.0.29](https://github.com/Esposter/Esposter/compare/v1.0.28...v1.0.29) (2024-05-18)
+
+**Note:** Version bump only for package parse-tmx
+
+
+
+
+
 ## [1.0.28](https://github.com/Esposter/Esposter/compare/v1.0.27...v1.0.28) (2024-05-18)
 
 **Note:** Version bump only for package parse-tmx